Minor fixes in weighted round robin algorithm implementation
* Fix typo in weights member
* Add source file to CMakeLists.txt
* Fix algorithm

+++ load_balancer/server/algorithms/weighted_round_robin.hpp
@@ -7,7 +7,7 @@ namespace eoanermine {
 namespace load_balancer {
 
 class WeightedRoundRobin : public RoundRobin {
-  std::vector<float> &weights;
+  std::vector<int> &weights;
 
 public:
   WeightedRoundRobin(std::vector<TargetInfo> &targets, std::vector<int>& weights);
